fix distributions data

+++ backend/domain/distribution.go
@@ -1,8 +1,8 @@
 package domain
 
 type DistributionData struct {
-	Hour       int   `json:"hour"`
-	Minute     int   `json:"minute"`
-	Mechanical int64 `json:"mechanical"`
-	Electric   int64 `json:"electric"`
+	Hour       int     `json:"hour"`
+	Minute     int     `json:"minute"`
+	Mechanical float64 `json:"mechanical"`
+	Electric   float64 `json:"electric"`
 }
